requirements.txt
Flask==2.0.1
RPi.GPIO==0.7.0
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
click==8.0.1 